util: Split up 'util' module into multiple files.
